Allow returning games from Find a Checkout
If an attendee brings a game back without their badge, or wants to
return a game for another attendee, we would have to scan the game, copy
the attendee's barcode, then navigate to the Checkout & Return page to
paste the barcode and return the game.

This change adds a Return button to the Find a Checkout page if the game
is currently checked out.